requirements.txt
asgiref
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==0.5.0
Django==3.2.*
idna==3.4
pyTelegramBotAPI
python-dotenv==0.20.0
pytz==2022.1
requests
sqlparse
urllib3==1.26.12
aiohttp
daphne
fastapi